style:애니메이션 파일 분리
